Add a skip command

+++ src/command/skip.h
@@ -2,6 +2,7 @@
 #define BRAGI_SKIP_H
 
 #include <dpp/dpp.h>
+#include "master.h"
 
 class Skip final {
 public:
@@ -10,7 +11,7 @@ public:
 	static void Exec(const dpp::message_create_t &event, const std::string &argument);
 
 private:
-	static dpp::message Exec(const dpp::snowflake &guild_id, const dpp::snowflake &channel_id, const int num_for_skip);
+	static dpp::message Exec(const dpp::snowflake &guild_id, const dpp::snowflake &channel_id, const int16 num_for_skip);
 };
 
 #endif
